% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max_volume.R
\name{calculate_max_volume}
\alias{calculate_max_volume}
\title{Calculate Maximum Volume}
\usage{
calculate_max_volume(hw_volume, sw_volume, maxvol = 38)
}
\arguments{
\item{hw_volume}{Numeric, hardwood volume.}

\item{sw_volume}{Numeric, softwood volume.}

\item{maxvol}{Numeric, Defaults to a biological maximum volume of 38 cords/acre.}
}
\value{
Numeric, the maximum standing volume in cords/acre.
}
\description{
This function calculates the maximum standing volume for a stand before the
SPA growth model defaults to a minimum growth rate.
}
\details{
The maximum standing volume is a function of the percentage of total
volume that is hardwood. This value is used as an input for calculating
the growth rate using the \code{calculate_growth_rate} function.

The biological maximum using defaults:
\itemize{
  \item 100\% HW - 38.0 cords/acre
  \item 75\% HW  - 39.9 cords/acre
  \item 50\% HW  - 41.8 cords/acre
  \item 25\% HW  - 43.7 cords/acre
  \item 0\% HW   - 45.6 cords/acre
}

While it is not required that the input volumes be in cords/acre, they
must be in the same units. The output will always be cords/acre.

The biological maximum is calculated as:
\deqn{maxvol \times (1 + (1 - ratio\_hw) \times 0.2)}
}
\examples{

hw_volume1 <- 18
sw_volume1 <- 22
calculate_max_volume(hw_volume1, sw_volume1)

# Collection of stands:

volumes_df <- data.frame(
  hw_volume = c(10, 20, 15, 0, 20, 0, 11),
  sw_volume = c(5, 10, 8, 2, 0, 20, 11)
)

volumes_df$max_vol <- apply(volumes_df, 1, function(row) {
  calculate_max_volume(row['hw_volume'], row['sw_volume'])
})

# Display Results
print(volumes_df)

volumes_df2 <- data.frame(
  hw_volume = c(10, 20, 15, 0, 20, 0, 11),
  sw_volume = c(5, 10, 8, 2, 0, 20, 11)
)

volumes_df2$max_vol <- apply(volumes_df2, 1, function(row) {
  calculate_max_volume(row['hw_volume'], row['sw_volume'], maxvol = 29)
})

# Display Results
print(volumes_df2)

}
\seealso{
[SPA7i::calculate_growth_rate]
}
\author{
Ryan Smith
}
\concept{Growth and Yield Functions}
