% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_table.R
\name{ci_table}
\alias{ci_table}
\title{Table of Confidence Intervals}
\usage{
ci_table(npv)
}
\arguments{
\item{npv}{A list containing the element `NPVs` from the `monteCarloAnalysis` function, or a numeric vector of potential NPV values.}
}
\value{
A data frame containing the confidence intervals (68, 80, 90 percent) and their respective lower and upper bounds.
}
\description{
This function generates a table of confidence intervals for a given vector of potential values.
The function calculates confidence intervals at the 68, 80, and 90 percent levels. The input can either
be a list generated by the `monteCarloAnalysis` function within the same package or a numeric vector
of potential values from an independent simulation.
}
\details{
If a list is passed, the function extracts the `NPVs` element from the list. If a numeric vector
is provided, it uses the values directly. The function first checks if the data is normally distributed
using the Shapiro-Wilk test. If the data is normally distributed (p-value > 0.05),
parametric confidence intervals are calculated based on the mean and standard deviation, using:

- 1 standard deviation for the 68 percent confidence interval
- 1.28 standard deviations for the 80 percent confidence interval
- 1.65 standard deviations for the 90 percent confidence interval

If the data is not normally distributed (p-value <= 0.05), non-parametric confidence intervals are calculated using quantiles:

- 68 percent confidence interval using the 16th and 84th percentiles
- 80 percent confidence interval using the 10th and 90th percentiles
- 90 percent confidence interval using the 5th and 95th percentiles
}
\examples{
# Example with a numeric vector from a Monte Carlo simulation (Normal Distribution)
set.seed(123)
npv_values <- rnorm(1000, mean = 200000, sd = 25000)
ci_table(npv_values)

# Example with a list from monteCarloAnalysis (Normal Distribution)
result <- list(NPVs = rnorm(1000, mean = 200000, sd = 25000),
  element1 = c(23, 23),
  element2 = c(32, 32)
  )
ci_table(result)

# Example with a numeric vector from an Exponential distribution (Not Normally Distributed)
set.seed(123)
npv_values_skewed <- rexp(1000, rate = 1/200000)
ci_table(npv_values_skewed)

}
\references{
Shapiro, S. S., & Wilk, M. B. (1965). An analysis of variance test for normality (complete samples). Biometrika, 52(3/4), 591-611.
}
\seealso{
\code{\link{monteCarloAnalysis}}

Other Summary Functions: 
\code{\link{generate_aac_summary}()},
\code{\link{generate_final_standing_summary}()},
\code{\link{generate_initial_matrix_summary}()}
}
\concept{Summary Functions}
