% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_prep.R
\name{data_preparation}
\alias{data_preparation}
\title{Load and prepare the data for analysis.}
\usage{
data_preparation(vols_path)
}
\arguments{
\item{vols_path}{A string representing the file path to the CSV file containing volume data.}
}
\value{
A cleaned dataframe with lowercase column names and the renamed 'matrix' column.
}
\description{
This function loads volume data from a CSV file, converts the column names to lowercase,
and renames any column containing the word "matrix" to "matrix", if it's not already named "matrix".
}
\examples{
\dontrun{
# Example 1: Column named 'lumpmatrix'
simulated_data_1 <- data.frame(
  lumpmatrix = c("matrix1", "matrix2", "matrix1"),
  product = c("perAcreHW", "perAcreSW", "perAcreHW"),
  cords = c(100, 50, 200),
  township = c("township1", "township1", "township2"),
  acres = c(10, 20, 15)
)
# Save to CSV in a temporary directory
temp_file_1 <- file.path(tempdir(), "SPA_test1.csv")
write.csv(simulated_data_1, temp_file_1, row.names = FALSE)

# Example usage where 'lumpmatrix' gets renamed to 'matrix'
vols_cleaned_1 <- data_preparation(temp_file_1)
print(vols_cleaned_1)

# Example 2: Column already named 'MATRIX'
simulated_data_2 <- data.frame(
  MATRIX = c("matrix1", "matrix2", "matrix1"),
  product = c("perAcreHW", "perAcreSW", "perAcreHW"),
  cords = c(100, 50, 200),
  township = c("township1", "township1", "township2"),
  acres = c(10, 20, 15)
)
# Save to CSV in a temporary directory
temp_file_2 <- file.path(tempdir(), "SPA_test2.csv")
write.csv(simulated_data_2, temp_file_2, row.names = FALSE)

# Example usage where 'MATRIX' is already 'matrix' and remains unchanged
vols_cleaned_2 <- data_preparation(temp_file_2)
print(vols_cleaned_2)
}
}
\seealso{
Other InputTransform Functions: 
\code{\link{calculateProductRatios}()},
\code{\link{mapWoodType}()}
}
\concept{InputTransform Functions}
