% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_prep.R
\name{dataPreparation}
\alias{dataPreparation}
\title{Load and prepare data for analysis from an Excel file.}
\usage{
dataPreparation(filepath)
}
\arguments{
\item{filepath}{A string representing the file path to the Excel file containing the data.}
}
\value{
A named list with two dataframes: "Inventory" and "Spatial".
The "Inventory" dataframe contains selected and cleaned data from the "Inventory" sheet.
The "Spatial" dataframe contains all lowercase text from the "Spatial" sheet.
}
\description{
This function processes an Excel file with two sheets:
"Inventory" and "Spatial". The "Inventory" sheet is cleaned by selecting
relevant columns (`township`, `matrix`, `acres`, `product`, `cords`), and any
column containing the word "matrix" is renamed to "matrix" if not already named so.
The "Spatial" sheet has all its colnames converted to lowercase.
}
\examples{
\dontrun{
# Example usage with an Excel file containing "Inventory" and "Spatial" sheets.
data_list <- dataPreparation("path_to_excel_file.xlsx")
inventory_data <- data_list$Inventory
spatial_data <- data_list$Spatial
print(inventory_data)
print(spatial_data)
}
}
\seealso{
Other InputTransform Functions: 
\code{\link{GrowthModelInput}()},
\code{\link{calculateProductRatios}()},
\code{\link{mapWoodType}()}
}
\concept{InputTransform Functions}
