% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthModelInputs.R
\name{GrowthModelInput}
\alias{GrowthModelInput}
\title{Prepare Growth Model Input Data.}
\usage{
GrowthModelInput(arc_df)
}
\arguments{
\item{arc_df}{The SPA export file from ArcGIS after loading with `data_preparation`.
A dataframe containing product data, including variables related to per-acre hardwood (HW) and softwood (SW) volumes.}
}
\value{
A dataframe with renamed columns for hardwood and softwood per-acre values, filtered to only include these columns, and pivoted. into the correct structure for further analysis.
}
\description{
This function prepares the input data for growth models by renaming the relevant columns
in the provided dataframe that correspond to per-acre hardwood and softwood values.
It ensures that columns related to per-acre volumes for hardwood and
softwood are renamed to `perAcreHW` and `perAcreSW` respectively, and filters the
dataframe to include only the relevant products.
The function then pivots the dataframe to set it up for AAC simulations.
}
\examples{
# Example dataframe
arc_df <- data.frame(
  township = c("township1", "township1", "township2", "township2"),
  matrix = c("matrix1", "matrix1", "matrix1", "matrix1"),
  product = c("PerAcreSW", "perAcreHW", "perAcreSW", "perAcre_HW"),
  cords = c(10, 15, 5, 8)
  )

# Generate input for growth models
final_results <- GrowthModelInput(arc_df)
print(final_results)

}
\seealso{
\code{\link{calculate_aac}}, \code{\link{run_aac_simulation}}

Other InputTransform Functions: 
\code{\link{calculateProductRatios}()},
\code{\link{dataPreparation}()},
\code{\link{mapWoodType}()}
}
\concept{InputTransform Functions}
