% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapWoodType.R
\name{mapWoodType}
\alias{mapWoodType}
\title{Map Product to Wood Type}
\usage{
mapWoodType(product)
}
\arguments{
\item{product}{A character vector containing product names.}
}
\value{
A character vector of the same length as `product` with the wood type ("HW", "SW", or "Other").
}
\description{
This function maps product names to their respective wood types, either "HW" for hardwood or "SW" for softwood.
Any product that is not recognized is mapped to "Other".
}
\details{
This function is designed to efficiently assign wood types based on a product list.
It uses predefined mappings for known products exported from AGM. If a product is not found in the
predefined list, it defaults to "Other".

The following product mappings are used:
\itemize{
  \item "ASl" and "ASp" - HW (Hardwood)
  \item "BFl" and "BFp" - SW (Softwood)
  \item "CEl" and "CEp" - SW
  \item "HVl", "HVp", "HVt" - HW
  \item "LVl", "LVp", "LVt" - HW
  \item "OSl" and "OSp" - SW
  \item "SPl" and "SPp" - SW
  \item "WPl" and "WPp" - SW
}
}
\examples{
products <- c("ASl", "BFl", "HVl", "unknown")
mapWoodType(products)

}
\seealso{
Other InputTransform Functions: 
\code{\link{calculateProductRatios}()},
\code{\link{data_preparation}()}
}
\concept{InputTransform Functions}
