% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProductVolumes.R
\name{plotProductVolumes}
\alias{plotProductVolumes}
\title{Plot Stocking, AAC, and Annual Species Group Standing Inventory}
\usage{
plotProductVolumes(summary, base_size = 14)
}
\arguments{
\item{summary}{A list containing annual volumes, summary volumes, and hardwood volumes from the `summarizeProductVolumes` function.}

\item{base_size}{An integer specifying the base font size for the plot. Defaults to 14.}
}
\value{
A list containing four ggplot objects:
1. `stockingplot`: Stocking summary of initial and final standing volumes for each species group.
2. `aacplot`: AAC summary of initial and final harvest volumes for each species group.
3. `annual_inventory_plot`: Line plot showing annual species group standing inventory.
4. `annual_aac_plot`: Line plot showing annual species group AAC.
}
\description{
This function takes the output of the `summarizeProductVolumes` function and creates four plots:
1. A stocking summary showing initial and final standing volumes for each species group.
2. An AAC (Annual Allowable Cut) summary showing initial and final harvest volumes for each species group.
3. A line plot showing the total standing inventory by year for each species group.
4. A line plot showing the AAC by year for each species group.

The number of years plotted will be equal to the number of years run in the `run_aac_for_all` function.

The species groups include Cedar, Hardwood, Other Softwood, Pine, Poplar, and Spruce-Fir.
The function automatically adjusts the y-axis to dynamically scale based on the maximum values in the data.
}
\examples{
\dontrun{
# Example usage:
summary <- summarizeProductVolumes(aac_results, ratios_df)
summary_plots <- plotProductVolumes(summary)
print(summary_plots$stockingplot)
print(summary_plots$aacplot)
print(summary_plots$annual_inventory_plot)
print(summary_plots$annual_inventory_plot)
}

}
\seealso{
\code{\link{summarizeProductVolumes}} for generating the summarized product volumes used as input to this function.

Other Reporting Functions: 
\code{\link{plotNPVDistribution}()},
\code{\link{summarizeProductVolumes}()}
}
\concept{Reporting Functions}
