% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPV.R
\name{simpleNPV}
\alias{simpleNPV}
\title{Net Present Value (NPV) Calculation}
\usage{
simpleNPV(Flow, Occurrence, Discount, Class = NULL)
}
\arguments{
\item{Flow}{Numeric vector. A series of cash flow values, where positive values represent revenue and negative values represent expenses.}

\item{Occurrence}{Numeric vector. The years from today when each cash flow occurs. Year 0 represents the present.}

\item{Discount}{Numeric. The discount rate to be applied to each cash flow, entered as a decimal (e.g., 0.05 for 5\%).}

\item{Class}{Optional character vector. Classify each flow as "Expense" or "Revenue", or "Negative" or "Positive". If missing, the function will assume that positive flows are revenue and negative flows are expenses.}
}
\value{
Numeric. The Net Present Value (NPV) of the cash flows based on the given discount rate.
}
\description{
This function calculates the Net Present Value (NPV) of a series of cash flows (positive or negative)
at a given discount rate. It automatically adjusts flows classified as "Expense" or "Negative" to
negative values if they are not already negative, and handles flows classified as "Revenue" or "Positive"
by ensuring they are positive. If no class is provided, it assumes that negative values are expenses and positive values are revenues.
}
\details{
The function computes the present value of each cash flow, which may represent either revenue or expense.
Cash flows classified as "Expense" or "Negative" are automatically converted to negative values if they are not already negative, and cash flows classified as "Revenue" or "Positive" are ensured to be positive.
If no `Class` vector is provided, the function assumes that positive flows are revenue and negative flows are expenses.
}
\examples{
Flow <- c(-400, -100, 60, 200, 6600)
Occurrence <- c(0, 5, 8, 15, 30)
Discount <- 0.06
simpleNPV(Flow, Occurrence, Discount)

\dontrun{
# Example with class vector
Flow <- c(400, 100, 60, 200, 6600)
Class <- c("Expense", "Expense", "Revenue", "Revenue", "Revenue")
Occurrence <- c(0, 5, 8, 15, 30)
Discount <- 0.06
simpleNPV(Flow, Occurrence, Discount, Class = Class)
}
}
\seealso{
Other Value Functions: 
\code{\link{landExpectVal}()},
\code{\link{presentValue}()}
}
\concept{Value Functions}
