% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeProductVolumes.R
\name{summarizeProductVolumes}
\alias{summarizeProductVolumes}
\title{Summarize Product Volumes for Initial and Final Years}
\usage{
summarizeProductVolumes(aac_results, ratios_df, years = 20)
}
\arguments{
\item{aac_results}{A dataframe containing AAC results for multiple years.}

\item{ratios_df}{A dataframe containing the ratios for hardwood and softwood for each township and matrix.}

\item{years}{Integer, number of years to calculate product volumes for (default is 20).}
}
\value{
A list containing three dataframes:
1. `annual_volumes`: Aggregated standing and harvest volumes for each species group across all years.
2. `summary_volumes`: Summarized standing and harvest volumes for the first and final years.
3. `hardwood_volumes`: Detailed breakdown of sawlog and pulp volumes and percentages for the hardwood species group.
}
\description{
This function takes the results of the product volume calculations and
aggregates the standing and harvest volumes for specified wood groupings
(hardwood, spruce-fir, cedar, pine, other softwood, poplar) for both initial and final years.
It also calculates the percentage of sawlogs and pulp for each year in the hardwood species group.

The function returns three dataframes:
1. `annual_volumes`: Summarized standing and harvest volumes for each species group across all years.
2. `summary_volumes`: Summarized standing and harvest volumes for the first and final years.
3. `hardwood_volumes`: Detailed breakdown of sawlog and pulp volumes and percentages for hardwood.
}
\examples{
\dontrun{
# Example usage:
summary_volumes <- summarizeProductVolumes(aac_results, ratios_df)
print(summary_volumes$annual_volumes)
print(summary_volumes$summary_volumes)
print(summary_volumes$hardwood_volumes)
}
}
\seealso{
\code{\link{calculateProductRatios}} and \code{\link{run_aac_for_all}} to create `aac_results` and `ratios_df`.

Other Reporting Functions: 
\code{\link{plotNPVDistribution}()},
\code{\link{plotProductVolumes}()}
}
\concept{Reporting Functions}
